+++ src/background/modules/apps/infrastructure.rs
@@ -33,7 +33,7 @@ pub fn get_focused_app() -> FocusedApp {
 #[tauri::command(async)]
 pub fn get_mouse_position() -> [i32; 2] {
     let point = Mouse::get_cursor_pos().unwrap_or_default();
-    [point.x(), point.y()]
+    [point.x, point.y]
 }
 
 #[tauri::command(async)]